moving viewmodel factory to its own file
