+++ app/src/main/java/com/example/mowpro/viewmodels/WeatherCardViewModel.kt
@@ -57,17 +57,3 @@ class WeatherCardViewModel(activity: AppCompatActivity) : ViewModel() {
         }
     }
 }
-
-class WeatherCardViewModelFactory(private val activity: AppCompatActivity) : ViewModelProvider.Factory {
-
-    private val logTag = "WeatherCardVMF"
-
-    override fun <T: ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(WeatherCardViewModel::class.java)) {
-            Log.d(logTag, "Factory created the ViewModel")
-            @Suppress("UNCHECKED_CAST")
-            return WeatherCardViewModel(activity = activity) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
