+++ app/src/main/java/com/example/mowpro/viewmodels/WeatherCardViewModelFactory.kt
@@ -1,0 +1,20 @@
+package com.example.mowpro.viewmodels
+
+import android.util.Log
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class WeatherCardViewModelFactory(private val activity: AppCompatActivity) : ViewModelProvider.Factory {
+
+    private val logTag = "WeatherCardVMF"
+
+    override fun <T: ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(WeatherCardViewModel::class.java)) {
+            Log.d(logTag, "Factory created the ViewModel")
+            @Suppress("UNCHECKED_CAST")
+            return WeatherCardViewModel(activity = activity) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
